import directly to Clickhouse database
Convert MySQL to ClickHouse SQL dialect + Implement ID auto-increment in CH database (#107)

- Dao layer SQL converted from MySQL compatibility into Clickhouse compatibility
- add  dependencies needed for Clickhouse driver (e.g. lz4)
- added internal id auto-increment class with on-exit persistence in database table
- incremental update logic also converted with adaptations and safeguards (due to loss of transactions)
    - one failing test requires merger of branch with transaction substitute (parallel work)
- CI testing converted using testcontainers library and docker deployment of clickhouse
- CI testing now resets database state before each test
- CI testing make sensitive to clickhouse/mysql differences (like case sensitivity or array order)
- introduce primary clickhouse table schema definition (analog to cgds.sql)
- refactor test code for standardization and reuse
- bugfixes where needed (e.g. incremental SV upload
- added geneset testing

+++ src/main/java/org/mskcc/cbio/portal/dao/DaoClinicalData.java
@@ -32,8 +32,6 @@
 
 package org.mskcc.cbio.portal.dao;
 
-import java.sql.*;
-import java.util.*;
 import org.apache.commons.lang3.StringUtils;
 import org.mskcc.cbio.portal.model.CancerStudy;
 import org.mskcc.cbio.portal.model.ClinicalAttribute;
@@ -42,6 +40,20 @@ import org.mskcc.cbio.portal.model.ClinicalParameterMap;
 import org.mskcc.cbio.portal.model.Patient;
 import org.mskcc.cbio.portal.model.Sample;
 import org.mskcc.cbio.portal.util.InternalIdUtil;
+
+import java.sql.Connection;
+import java.sql.PreparedStatement;
+import java.sql.ResultSet;
+import java.sql.SQLException;
+import java.util.ArrayList;
+import java.util.Collection;
+import java.util.Collections;
+import java.util.HashMap;
+import java.util.HashSet;
+import java.util.LinkedHashMap;
+import java.util.List;
+import java.util.Map;
+import java.util.Set;
 
 /**
  * Data Access Object for `clinical` table
@@ -53,12 +65,12 @@ public final class DaoClinicalData {
     public static final String SAMPLE_ATTRIBUTES_TABLE = "clinical_sample";
     public static final String PATIENT_ATTRIBUTES_TABLE = "clinical_patient";
 
-    private static final String SAMPLE_ATTRIBUTES_INSERT = "INSERT INTO " + SAMPLE_ATTRIBUTES_TABLE + "(`INTERNAL_ID`,`ATTR_ID`,`ATTR_VALUE`) VALUES(?,?,?)";
-    private static final String PATIENT_ATTRIBUTES_INSERT = "INSERT INTO " + PATIENT_ATTRIBUTES_TABLE + "(`INTERNAL_ID`,`ATTR_ID`,`ATTR_VALUE`) VALUES(?,?,?)";
+    private static final String SAMPLE_ATTRIBUTES_INSERT = "INSERT INTO " + SAMPLE_ATTRIBUTES_TABLE + "(`internal_id`,`attr_id`,`attr_value`) VALUES(?,?,?)";
+    private static final String PATIENT_ATTRIBUTES_INSERT = "INSERT INTO " + PATIENT_ATTRIBUTES_TABLE + "(`internal_id`,`attr_id`,`attr_value`) VALUES(?,?,?)";
 
-    private static final String SAMPLE_ATTRIBUTES_DELETE = "DELETE FROM " + SAMPLE_ATTRIBUTES_TABLE + " WHERE `INTERNAL_ID` = ?";
+    private static final String SAMPLE_ATTRIBUTES_DELETE = "DELETE FROM " + SAMPLE_ATTRIBUTES_TABLE + " WHERE `internal_id` = ?";
 
-    private static final String PATIENT_ATTRIBUTES_DELETE = "DELETE FROM " + PATIENT_ATTRIBUTES_TABLE + " WHERE `INTERNAL_ID` = ?";
+    private static final String PATIENT_ATTRIBUTES_DELETE = "DELETE FROM " + PATIENT_ATTRIBUTES_TABLE + " WHERE `internal_id` = ?";
     private static final Map<String, String> sampleAttributes = new HashMap<String, String>();
     private static final Map<String, String> patientAttributes = new HashMap<String, String>();
 
@@ -87,7 +99,7 @@ public final class DaoClinicalData {
             pstmt = con.prepareStatement("SELECT * FROM " + table);
             rs = pstmt.executeQuery();
             while (rs.next()) {
-                cache.put(rs.getString("ATTR_ID"), rs.getString("ATTR_ID"));
+                cache.put(rs.getString("attr_id"), rs.getString("attr_id"));
             }
         }
         catch (SQLException e) {
@@ -113,8 +125,8 @@ public final class DaoClinicalData {
     public static int addDatum(String query, String tableName,
                                int internalId, String attrId, String attrVal) throws DaoException
     {
-        if (MySQLbulkLoader.isBulkLoad()) {
-            MySQLbulkLoader.getMySQLbulkLoader(tableName).insertRecord(Integer.toString(internalId),
+        if (ClickHouseBulkLoader.isBulkLoad()) {
+            ClickHouseBulkLoader.getClickHouseBulkLoader(tableName).insertRecord(Integer.toString(internalId),
                 attrId,
                 attrVal);
             return 1;
@@ -191,7 +203,7 @@ public final class DaoClinicalData {
             con = JdbcUtil.getDbConnection(DaoClinicalData.class);
 
             pstmt = con.prepareStatement("SELECT * FROM " + table +
-                " WHERE INTERNAL_ID=? AND ATTR_ID=?");
+                " WHERE internal_id=? AND attr_id=?");
             pstmt.setInt(1, internalId);
             pstmt.setString(2, attrId);
 
@@ -225,7 +237,7 @@ public final class DaoClinicalData {
         ResultSet rs = null;
 
         List<ClinicalData> clinicals = new ArrayList<ClinicalData>();
-        String sql = ("SELECT * FROM " + table + " WHERE `INTERNAL_ID` IN " +
+        String sql = ("SELECT * FROM " + table + " WHERE `internal_id` IN " +
             "(" + generateIdsSql(internalIds) + ")");
 
         try {
@@ -404,9 +416,12 @@ public final class DaoClinicalData {
         Connection con = null;
         PreparedStatement pstmt = null;
         try {
+            if (sampleInternalIds.isEmpty()) {
+                return;
+            }
             con = JdbcUtil.getDbConnection(DaoClinicalData.class);
             pstmt = con.prepareStatement("DELETE FROM " + SAMPLE_ATTRIBUTES_TABLE
-                    + " WHERE `ATTR_ID` = ? AND `INTERNAL_ID` IN ("
+                    + " WHERE `attr_id` = ? AND `internal_id` IN ("
                     + String.join(",", Collections.nCopies(sampleInternalIds.size(), "?"))
                     + ")");
             int parameterIndex = 1;
@@ -452,9 +467,9 @@ public final class DaoClinicalData {
 
         List<ClinicalData> clinicals = new ArrayList<ClinicalData>();
 
-        String sql = ("SELECT * FROM " + table + " WHERE `INTERNAL_ID` IN " +
+        String sql = ("SELECT * FROM " + table + " WHERE `internal_id` IN " +
             "(" + generateIdsSql(internalIds) + ") " +
-            " AND ATTR_ID IN ('"+ StringUtils.join(attributeIds, "','")+"') ");
+            " AND attr_id IN ('"+ StringUtils.join(attributeIds, "','")+"') ");
 
         try {
             con = JdbcUtil.getDbConnection(DaoClinicalData.class);
@@ -484,13 +499,13 @@ public final class DaoClinicalData {
             con = JdbcUtil.getDbConnection(DaoClinicalData.class);
 
             pstmt = con.prepareStatement("SELECT * FROM clinical_patient WHERE" +
-                " ATTR_ID IN ('" + StringUtils.join(attributeIds, "','") +"') ");
+                " attr_id IN ('" + StringUtils.join(attributeIds, "','") +"') ");
 
             List<Integer> patients = getPatientIdsByCancerStudy(internalCancerStudyId);
 
             rs = pstmt.executeQuery();
             while(rs.next()) {
-                Integer patientId = rs.getInt("INTERNAL_ID");
+                Integer patientId = rs.getInt("internal_id");
                 if (patients.contains(patientId)) {
                     clinicals.add(extract(PATIENT_ATTRIBUTES_TABLE, internalCancerStudyId, rs));
                 }
@@ -508,11 +523,11 @@ public final class DaoClinicalData {
 
     private static ClinicalData extract(String table, int internalCancerStudyId, ResultSet rs) throws SQLException {
         // get 
-        String stableId = getStableIdFromInternalId(table, rs.getInt("INTERNAL_ID"));
+        String stableId = getStableIdFromInternalId(table, rs.getInt("internal_id"));
         return new ClinicalData(internalCancerStudyId,
             stableId,
-            rs.getString("ATTR_ID"),
-            rs.getString("ATTR_VALUE"));
+            rs.getString("attr_id"),
+            rs.getString("attr_value"));
     }
 
     private static String getStableIdFromInternalId(String table, int internalId)
@@ -664,8 +679,8 @@ public final class DaoClinicalData {
 
         try{
             con = JdbcUtil.getDbConnection(DaoClinicalData.class);
-            pstmt = con.prepareStatement ("SELECT INTERNAL_ID FROM `" + tableName + "`"
-                + " WHERE ATTR_ID=? AND ATTR_VALUE=?");
+            pstmt = con.prepareStatement ("SELECT internal_id FROM `" + tableName + "`"
+                + " WHERE attr_id=? AND attr_value=?");
             pstmt.setString(1, paramName);
             pstmt.setString(2, paramValue);
             rs = pstmt.executeQuery();
@@ -674,7 +689,7 @@ public final class DaoClinicalData {
 
             while (rs.next())
             {
-                ids.add(rs.getInt("INTERNAL_ID"));
+                ids.add(rs.getInt("internal_id"));
             }
 
             return ids;
@@ -695,10 +710,10 @@ public final class DaoClinicalData {
         try{
             con = JdbcUtil.getDbConnection(DaoClinicalData.class);
             pstmt = con.prepareStatement("select " +
-                "distinct ATTR_VALUE as attributeValue, " +
-                "ATTR_ID as attributeID from clinical_sample " +
-                "where ATTR_ID in (?, ?) and INTERNAL_ID in (" +
-                "select INTERNAL_ID from sample where STABLE_ID in ('"
+                "distinct attr_value as attributeValue, " +
+                "attr_id as attributeID from clinical_sample " +
+                "where attr_id in (?, ?) and internal_id in (" +
+                "select internal_id from sample where stable_id in ('"
                 + StringUtils.join(samplesList,"','")+"'))");
             pstmt.setString(1, ClinicalAttribute.CANCER_TYPE);
             pstmt.setString(2, ClinicalAttribute.CANCER_TYPE_DETAILED);
@@ -737,5 +752,46 @@ public final class DaoClinicalData {
         finally {
             JdbcUtil.closeAll(DaoClinicalData.class, con, pstmt, rs);
         }
+    }
+
+    private static List<Map.Entry<Integer, String>> getExistingAttributes(String tableName, List<Map.Entry<Integer, String>> internalPatientIdAttributes) throws DaoException {
+        if (internalPatientIdAttributes.isEmpty()) {
+            return Collections.emptyList();
+        }
+        Connection con = null;
+        PreparedStatement pstmt = null;
+        ResultSet rs = null;
+
+        try {
+            con = JdbcUtil.getDbConnection(DaoClinicalData.class);
+            pstmt = con.prepareStatement("SELECT internal_id, attr_id FROM `"
+                    + tableName + "` WHERE (internal_id, attr_id) IN ("
+                    + String.join(",", Collections.nCopies(internalPatientIdAttributes.size(), "(?, ?)")) + ")");
+            int parameterIndex = 1;
+            for (Map.Entry<Integer, String> entry : internalPatientIdAttributes) {
+                pstmt.setInt(parameterIndex++, entry.getKey());
+                pstmt.setString(parameterIndex++, entry.getValue());
+            }
+            rs = pstmt.executeQuery();
+
+            List<Map.Entry<Integer, String>> result = new ArrayList<>();
+            while (rs.next()) {
+                result.add(Map.entry(rs.getInt("internal_id"), rs.getString("attr_id")));
+            }
+
+            return result;
+        } catch (SQLException e) {
+            throw new DaoException(e);
+        } finally {
+            JdbcUtil.closeAll(DaoClinicalData.class, con, pstmt, rs);
+        }
+    }
+
+    public static List<Map.Entry<Integer, String>> getExistingPatientAttributes(List<Map.Entry<Integer, String>> internalPatientIdAttributes) throws DaoException {
+        return getExistingAttributes(PATIENT_ATTRIBUTES_TABLE, internalPatientIdAttributes);
+    }
+
+    public static List<Map.Entry<Integer, String>> getExistingSampleAttributes(List<Map.Entry<Integer, String>> internalPatientIdAttributes) throws DaoException {
+        return getExistingAttributes(SAMPLE_ATTRIBUTES_TABLE, internalPatientIdAttributes);
     }
 }
